refactor(getGlobalFocusStyles): moved to internal
